Moved home page endpoint
